backend/requirements.txt
python-dotenv==1.0.0
livekit==0.17.0
livekit-agents==0.9.0
livekit-plugins-openai==0.7.0
livekit-plugins-deepgram==0.6.0
livekit-plugins-cartesia==0.2.0
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
aiohttp==3.10.5
openai>=1.3.0
requests>=2.31.0